+++ Cargo.toml
@@ -22,6 +22,7 @@ members = [
     "util/dot-movement",
     "util/flocks",
     "util/godfig",
+    "util/movement-algs",
     "util/movement-types",
     "util/tracing",
     "networks/suzuka/*",
@@ -76,6 +77,7 @@ mcr-settlement-config = { path = "protocol-units/settlement/mcr/config" }
 mcr-settlement-manager = { path = "protocol-units/settlement/mcr/manager" }
 mcr-settlement-setup = { path = "protocol-units/settlement/mcr/setup" }
 ## types
+movement-algs = { path = "util/movement-algs" }
 movement-types = { path = "util/movement-types" }
 ## dot movement
 dot-movement = { path = "util/dot-movement" }
