--- util/movement-types/src/algs/congestion_control/mod.rs
@@ -1,1 +1,0 @@
-pub mod split;
