--- util/movement-types/src/algs/congestion_control/split.rs
@@ -1,4 +1,0 @@
-pub trait Splitting<T>
-    where T: Sized {
-    fn split(self) -> Vec<T>;
-}
