--- util/movement-types/src/algs/mod.rs
@@ -1,2 +1,0 @@
-pub mod congestion_control;
-pub mod grouping_heuristic;
